refactor: move useBookmarkFolders() to repository
